fixed build issues, removed cyclic dependency
